# 791 add performance data in codegen rebase