--- a/Codegen/src/Generators/SignaturesGenerator.cs
+++ b/Codegen/src/Generators/SignaturesGenerator.cs
@@ -107,7 +107,7 @@ namespace TypeCobol.Codegen.Generators
 
         public List<Diagnostic> Diagnostics { get; }
         public IReadOnlyDictionary<string, TimeSpan> PerformanceReport { get; private set; }
-        public string TypeCobolVersion { get; set; }
+        public string TypeCobolVersion { get; }
 
         public bool HasLineMapData => false;
     }