Rebased version of #1446 
Time measurements for Codegen. Example of output :
```
FILE;ITERATION;STEP;TIME_TAKEN
CGM110.cbl;1;CreateTargetDocument;145,6953
CGM110.cbl;1;Qualifier;152,3024
CGM110.cbl;1;Actions;765,8847
CGM110.cbl;1;TreeToCode;5281,6669
CGM110.cbl;2;CreateTargetDocument;154,7135
CGM110.cbl;2;Qualifier;0,0008
CGM110.cbl;2;Actions;133,9764
CGM110.cbl;2;TreeToCode;4751,2268

Time per step :
TreeToCode               (4) : Total =      10032,8937 ms, Avg =      5016,44685 ms
Actions                  (3) : Total =        899,8611 ms, Avg =       449,93055 ms
CreateTargetDocument     (1) : Total =        300,4088 ms, Avg =        150,2044 ms
Qualifier                (2) : Total =        152,3032 ms, Avg =         76,1516 ms

Global Codegen time : Total = 11385,4668 ms, Avg = 2846,3667 ms
```

The first block can be copy-pasted as CSV, second block is more human-readable.

Also fixes bug in LinearGeneration occuring when last lines of a input files are comments.

**Notes**
- Performance tests are Ignored by default
- Test_BigTypes_1Procedure is KO due to a bug (see #1445)
- Test_GlobalStorage is KO due to a bug in CrossChecker but fixed in #1081, see commit aa27f63e17281fed29081ef8a611a871ec4de356