@smedilol 2019-10-18
With this output:
``` 
Time per step :
TreeToCode               (4) : Total =      10032,8937 ms, Avg =      5016,44685 ms
Actions                  (3) : Total =        899,8611 ms, Avg =       449,93055 ms
CreateTargetDocument     (1) : Total =        300,4088 ms, Avg =        150,2044 ms
Qualifier                (2) : Total =        152,3032 ms, Avg =         76,1516 ms

Global Codegen time : Total = 11385,4668 ms, Avg = 2846,3667 ms
```

Is it possible to sort step by their index rather than by the duration.
It'd be easier to compare the report of 2 performance tests.

Can you also the percentage of the duration for each step ?